contents of etc/etcskel-version to be updated with `date -I` (gnu date) when wanting to trigger an update for subscribers.
subscribers will run bin/etcskel-checkversion via cron.daily, which
compares the local etcskel-version with the one on the git server.  if
the local is up-to-date, etcskel-checkversion will exit silently, else
it will output info about the version difference & a reminder to run
etcskel-update, which cron will then email to the configured address.